#2679 Bind HibernateTypeMappings through a provider
