+++ platform-dao/src/test/java/ua/com/fielden/platform/meta/UnionEntityMetadataTest.java
@@ -6,7 +6,7 @@ import ua.com.fielden.platform.entity.query.IDbVersionProvider;
 import ua.com.fielden.platform.meta.Assertions.EntityA;
 import ua.com.fielden.platform.meta.PropertyMetadata.Calculated;
 import ua.com.fielden.platform.meta.PropertyTypeMetadata.Primitive;
-import ua.com.fielden.platform.persistence.types.PlatformHibernateTypeMappings;
+import ua.com.fielden.platform.persistence.types.PlatformHibernateTypeMappingsProvider;
 import ua.com.fielden.platform.sample.domain.TgBogieLocation;
 import ua.com.fielden.platform.sample.domain.TgFuelType;
 
@@ -17,7 +17,7 @@ import static ua.com.fielden.platform.entity.query.IDbVersionProvider.constantDb
 public class UnionEntityMetadataTest {
 
     private static final IDbVersionProvider dbVersionProvider = constantDbVersion(DbVersion.MSSQL);
-    private static final TestDomainMetadataGenerator generator = TestDomainMetadataGenerator.wrap(new DomainMetadataGenerator(new PlatformHibernateTypeMappings.Provider(dbVersionProvider).get(), dbVersionProvider));
+    private static final TestDomainMetadataGenerator generator = TestDomainMetadataGenerator.wrap(new DomainMetadataGenerator(new PlatformHibernateTypeMappingsProvider(dbVersionProvider).get(), dbVersionProvider));
 
     @Test
     public void union_entity_gets_union_nature() {
